Add content types Home and Offering
Add folders for both
Add an empty home.json file in data/home
Add index.tsx files for both, with ts types

TODO: Not sure if the type definitions are correct...
lets see how the json looks for reference widgets

#Note:
With a files collections, the CMS expects each file to exist already.
It will not create the file if it does not exist. It is fine if the file
is empty.