Use the new CMS fields in the landing page component & its dependencies
#Reservations
I had to use a lot of prop drilling.

#Alternatives
We could avoid the prop drilling by using the useHome hook in the leaf components
This would make it less effort to rename fields in the future

#Problems with the alternative
Storybook issues: the components' appearance would no longer depend just
on props, but instead on props and CMS data. This is not desired when
a component is displayed in the storybook

#Possible ways to get best of both worlds
- Make useHome read data from context. Wrap production code in a context
containing the constant, in-CMS data. Wrap storybook code in a different
context for each story, that provides example data.

- Flatten the dependency hierachy of the landing page