Try to use importAll
Problem: The Projects data is imported twice: In src/data/project/index.tsx
and once again independently in gatsby-node.js!

Problem 2: Gatsby-node.js does not support require.context since that is a
webapck feature. May be possible to fix this issue by changing the
gatsby webpack configuration.

Once that is done it may be possible to just call getProjects("en") or
getProjects("de") in gatsby-node.js